requirements.txt
# Core
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9

# AI & LLM
langchain>=0.2.0
langgraph>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
chromadb>=0.5.0


# Async & Task Queue
celery>=5.3.0
django-celery-beat>=2.6.0
redis>=5.0.0


# Optional but helpful
markdown
requests
